+++ migrations/2025-07-22-032815_create-players-table/down.sql
@@ -1,3 +1,18 @@
+drop trigger on_insert_player_recomposition_trigger on data.player_recompositions;
+drop function data.on_insert_player_recomposition;
+drop table data.player_recompositions;
+
+drop trigger on_insert_player_paradigm_shift_trigger on data.player_paradigm_shifts;
+drop function data.on_insert_player_paradigm_shift;
+drop table data.player_paradigm_shifts;
+
+drop trigger on_insert_player_augment_trigger on data.player_augments;
+drop function data.on_insert_player_augment;
+drop table data.player_augments;
+
+drop table taxa.attribute;
+drop table taxa.attribute_category;
+
 drop trigger on_insert_player_version_trigger on data.player_versions;
 drop function data.on_insert_player_version;
 drop table data.player_versions;